+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Rotate different sites for availability
 - Show episode 0 if have
 - Add warm cache for popular catalogs
+- Fix client cache for all resource route
 
 # v0.3.0 2026-03-07